Updates, mangler kun Orderline som skal på plads. Det er dog i SQL

+++ Service-Api/DTOs/OrdersDto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int OrderNumber { get; set; }
         public DateTime DateTime { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ShopId { get; set; }
         //Empty Constructor
         public OrdersDto() { }
