+++ ServiceData/ModelLayer/Orders.cs
@@ -11,13 +11,13 @@ namespace ServiceData.ModelLayer
         public int Id { get; set; }
         public int OrderNumber { get; set; }
         public DateTime DateTime { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ShopId { get; set; }
         //Empty Constructor
         public Orders() { }
 
         //Constructor with parameters
-        public Orders( int orderNumber, DateTime dateTime, double totalPrice, int shopId)
+        public Orders( int orderNumber, DateTime dateTime, decimal totalPrice, int shopId)
         {
             
             OrderNumber = orderNumber;
@@ -27,7 +27,7 @@ namespace ServiceData.ModelLayer
         }
 
         //Constructor with Id
-        public Orders(int id, int orderNumber, DateTime dateTime, double totalPrice, int shopId) : this (orderNumber, dateTime, totalPrice, shopId)
+        public Orders(int id, int orderNumber, DateTime dateTime, decimal totalPrice, int shopId) : this (orderNumber, dateTime, totalPrice, shopId)
         {
             Id = id;
         }
